pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finanalytics-ai"
version = "0.1.0"
description = "FinAnalytics AI — Framework de Análise e Busca de Investimentos"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "FinAnalytics AI Team" }]

dependencies = [
    # Async HTTP
    "aiohttp>=3.9",
    "httpx>=0.27",
    # Database (async)
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    # Data validation & settings
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    # Structured logging
    "structlog>=24.1",
    # Retry / resilience
    "tenacity>=8.3",
    # Cache / rate limiting
    "redis>=5.0",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # PDF reports
    "reportlab>=4.0",
    # Observability
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "prometheus-client>=0.20",
    # Numerics & time series
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.13",
    # Environment
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "mypy>=1.10",
    "ruff>=0.4",
    "black>=24.4",
    "sqlalchemy[mypy]",
]

[project.scripts]
finanalytics = "finanalytics_ai.interfaces.cli.commands:main"

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
    "--cov=src/finanalytics_ai",
    "--cov-report=term-missing",
    "--cov-report=html:reports/coverage",
]
markers = [
    "unit: fast, isolated unit tests",
    "integration: tests that hit real I/O (db, network)",
    "slow: tests that take more than 1s",
]

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
# src layout — tell mypy where to look
mypy_path = "src"
packages = ["finanalytics_ai"]
# Plugins
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

# Relax only for third-party stubs that aren't typed
[[tool.mypy.overrides]]
module = ["aiohttp.*", "asyncpg.*", "tenacity.*", "scipy.*", "pandas.*"]
ignore_missing_imports = true

# ── ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
src = ["src"]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "SIM", "TCH", "ANN", "PT", "RUF"]
ignore = [
    "ANN401",  # Any e' legitimo em wrappers, pools asyncpg e adaptadores
    "B904",    # raise X from err - 81 ocorrencias em routes/infra; backlog incremental
    "RUF001",  # x, -, sigma em strings (notacao matematica intencional)
    "RUF002",  # idem em docstrings
    "RUF003",  # idem em comentarios
]

[tool.ruff.lint.per-file-ignores]
"src/finanalytics_ai/interfaces/api/**" = ["B008", "ANN001", "ANN201", "ANN202"]
"src/finanalytics_ai/application/services/fixed_income_service.py" = ["E402"]
"src/finanalytics_ai/application/services/fund_analysis_service.py" = ["E501"]
"src/finanalytics_ai/application/services/rf_portfolio_service.py" = ["E402", "ANN202"]
"src/finanalytics_ai/application/services/ohlc_1m_service.py" = ["ANN001", "ANN202"]
"src/finanalytics_ai/application/services/patrimony_service.py" = ["ANN001"]
"src/finanalytics_ai/infrastructure/adapters/market_data_client.py" = ["ANN001", "ANN201"]
"src/finanalytics_ai/infrastructure/database/repositories/ticker_service.py" = ["ANN001", "ANN202"]
"src/finanalytics_ai/interfaces/api/routes/fixed_income.py" = ["E402"]
"src/finanalytics_ai/interfaces/api/routes/backtest.py" = ["E402"]
"src/finanalytics_ai/interfaces/api/run.py" = ["E402"]
"src/finanalytics_ai/workers/**" = ["ANN001", "ANN202"]
"tests/**" = ["ANN001", "ANN201", "ANN202", "ANN401"]

[tool.ruff.lint.isort]
known-first-party = ["finanalytics_ai"]

[tool.ruff.format]
# Make ruff format compatible with black
quote-style = "double"
indent-style = "space"

# ── black ─────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 110
target-version = ["py312"]

# ── coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/finanalytics_ai"]
omit = ["*/migrations/*", "*/__init__.py"]

[tool.coverage.report]
show_missing = true
fail_under = 80

# ── Interfaces — FastAPI usa decoradores dinâmicos; strict aqui gera ruído ────
[[tool.mypy.overrides]]
module = [
    "finanalytics_ai.interfaces.api.routes.*",
    "finanalytics_ai.interfaces.api.app",
    "finanalytics_ai.interfaces.api.run",
]
disable_error_code = ["no-untyped-def", "no-untyped-call", "type-arg", "no-any-return"]

# ── Infrastructure / DB — SQLAlchemy 1.x Column[X] vs mapped_column (SA 2.0) ─
# Trade-off: migrar para SA 2.0 mapped_column resolve isso definitivamente;
# por ora suprimimos arg-type e assignment nas repos para não bloquear o CI.
[[tool.mypy.overrides]]
module = [
    "finanalytics_ai.infrastructure.database.repositories.*",
    "finanalytics_ai.infrastructure.database.models.*",
]
disable_error_code = ["arg-type", "assignment", "type-arg", "return-value"]

# ── Infrastructure / Adapters + Cache — stubs externos incompletos ────────────
[[tool.mypy.overrides]]
module = [
    "finanalytics_ai.infrastructure.adapters.*",
    "finanalytics_ai.infrastructure.cache.*",
    "finanalytics_ai.infrastructure.auth.*",
    "finanalytics_ai.infrastructure.queue.*",
    "finanalytics_ai.infrastructure.timescale.*",
    "finanalytics_ai.infrastructure.reports.*",
    "finanalytics_ai.infrastructure.notifications.*",
]
disable_error_code = ["no-untyped-def", "no-redef", "type-arg", "no-any-return", "unused-ignore"]

# ── Workers — scripts operacionais; baixa prioridade de tipagem ───────────────
[[tool.mypy.overrides]]
module = ["finanalytics_ai.workers.*"]
disable_error_code = ["no-untyped-def", "type-arg", "no-any-return", "return-value"]

# ── Application services — relaxar apenas type-arg (dict sem params) ──────────
[[tool.mypy.overrides]]
module = ["finanalytics_ai.application.services.*"]
disable_error_code = ["type-arg"]

# ── hatch (build) ─────────────────────────────────────────────────────────────
[tool.hatch.build.targets.wheel]
packages = ["src/finanalytics_ai"]

# Sprint 2 additions (append to [project] dependencies manually)
# fastapi>=0.111, uvicorn[standard]>=0.30, alembic>=1.13

